parser: move nested_parenthesis() helper into node module
