+++ askama_parser/src/lib.rs
@@ -202,54 +202,6 @@ fn char_lit(i: &str) -> IResult<&str, &str> {
     Ok((i, s.unwrap_or_default()))
 }
 
-fn nested_parenthesis(i: &str) -> IResult<&str, ()> {
-    let mut nested = 0;
-    let mut last = 0;
-    let mut in_str = false;
-    let mut escaped = false;
-
-    for (i, b) in i.chars().enumerate() {
-        if !(b == '(' || b == ')') || !in_str {
-            match b {
-                '(' => nested += 1,
-                ')' => {
-                    if nested == 0 {
-                        last = i;
-                        break;
-                    }
-                    nested -= 1;
-                }
-                '"' => {
-                    if in_str {
-                        if !escaped {
-                            in_str = false;
-                        }
-                    } else {
-                        in_str = true;
-                    }
-                }
-                '\\' => {
-                    escaped = !escaped;
-                }
-                _ => (),
-            }
-        }
-
-        if escaped && b != '\\' {
-            escaped = false;
-        }
-    }
-
-    if nested == 0 {
-        Ok((&i[last..], ()))
-    } else {
-        Err(nom::Err::Error(error_position!(
-            i,
-            ErrorKind::SeparatedNonEmptyList
-        )))
-    }
-}
-
 fn path(i: &str) -> IResult<&str, Vec<&str>> {
     let root = opt(value("", ws(tag("::"))));
     let tail = separated_list1(ws(tag("::")), identifier);
